requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.2
playwright==1.55.0
PyYAML==6.0.3
requests==2.32.5

